# backend/parabellum_swagger.yaml
openapi: 3.0.0
info:
  title: API de Gestion Parabellum Group
  version: "1.0.0"
  description: Documentation complète de l'API pour l'application de gestion du Parabellum Group, incluant toutes les méthodes HTTP.
  contact:
    name: Support Parabellum
    email: support@parabellum.com
servers:
  - url: http://localhost:3001/api # Ajuster le port si nécessaire
    description: Serveur de développement
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Role:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        nom:
          type: string
          example: "Administrateur"
        description:
          type: string
          nullable: true
          example: "Accès complet à toutes les fonctionnalités"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          example: "Description de l'erreur"
        details:
          type: object
          nullable: true
          example:
            field: "Message spécifique au champ"
security:
  - bearerAuth: []
paths:
...
  /roles:
    get:
      tags:
        - Roles
      summary: Récupérer la liste de tous les rôles
      operationId: getAllRoles
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Liste des rôles récupérée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
              examples:
                success:
                  value:
                    - id: 1
                      nom: "Administrateur"
                      description: "Accès complet"
                    - id: 2
                      nom: "Technicien"
                      description: "Accès aux missions et interventions"
        '401':
          description: Non autorisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  value:
                    error: true
                    message: "Token d'authentification manquant ou invalide."
        '500':
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                serverError:
                  value:
                    error: true
                    message: "Erreur interne du serveur."
    # ... autres méthodes pour /roles et /roles/{id} avec exemples ...
    post:
      tags:
        - Roles
      summary: Créer un nouveau rôle
      # ... (contenu similaire à Clients post avec exemples)
      responses:
        '201':
          description: Rôle créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
              examples:
                success:
                  value:
                    id: 3
                    nom: "Visiteur"
                    description: "Accès limité en lecture seule"
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validationError:
                  value:
                    error: true
                    message: "Le nom du rôle est requis."
...